+++ src/hooksFactory.js
@@ -4,9 +4,12 @@ import { useSelector, useDispatch } from 'react-redux';
 export default (objectName, { byKey, parent }, {
   mapToProps,
   mapToPropsStripped,
-  actions,
-  actionsStripped,
-  actionsStrippedToFullName,
+  asyncActions,
+  syncActions,
+  asyncActionsStripped,
+  syncActionsStripped,
+  asyncActionsIncluded,
+  syncActionsIncluded,
 }) => {
   const getUseFactory = stripped => ({ [byKey]: id, [parent]: parentId, ...restProps } = {}) => {
     if (!parent && typeof parentId !== 'undefined') {
@@ -22,57 +25,62 @@ export default (objectName, { byKey, parent }, {
         restProps
       );
     }
+
     // Reusable object when parentId is defined
     const idObj = typeof id !== 'undefined' ? { [byKey]: id } : {}
     const parentObj = parent && typeof parentId !== 'undefined' ? { [parent]: parentId } : {}
+    
     // Reuse mapToProps functions to get state
     const obj = useSelector(state => 
       stripped
         ? mapToPropsStripped(state, { ...idObj, ...parentObj })
         : mapToProps(state, { ...idObj, ...parentObj })
     );
-    const dispatch = useDispatch()
+    const dispatch = useDispatch();
 
     return {
       ...obj,
-      ...Object.entries(actionsStrippedToFullName).reduce(
-        (o, [strippedName, fullName]) => {
+      ...Object.entries(stripped ? asyncActionsStripped : asyncActions).reduce(
+        (o, [actionName, actionFunction]) => {
           let dispatchableAction
-          switch (strippedName) {
-            case 'create':
-              dispatchableAction = (obj = {}, ...restArgs) =>
-                dispatch(actionsStripped[strippedName]({ ...parentObj, ...obj }, ...restArgs)
-              );
-            case 'update':
-              dispatchableAction = (obj = {}, ...restArgs) =>
-                dispatch(actionsStripped[strippedName]({ ...idObj, ...parentObj, ...obj }, ...restArgs)
-              );
-              break
-            case 'delete':
-              dispatchableAction = (obj = {}, ...restArgs) =>
-                dispatch(actionsStripped[strippedName]({
-                  ...idObj,
-                  ...parentObj,
-                  ...typeof obj === 'object' ? obj : { byKey: obj },
-                }, ...restArgs)
-              );
-              break
-            default:
-              dispatchableAction = (...args) => dispatch(actionsStripped[strippedName](...args));
+          if (actionName.startsWith('create')) {
+            dispatchableAction = (obj = {}, ...restArgs) =>
+              dispatch(actionFunction({ ...parentObj, ...obj }, ...restArgs)
+            );
+          } else if (actionName.startsWith('update')) {
+            dispatchableAction = (obj = {}, ...restArgs) =>
+              dispatch(actionFunction({ ...idObj, ...parentObj, ...obj }, ...restArgs)
+            );
+          } else if (actionName.startsWith('delete')) {
+            dispatchableAction = (obj = {}, ...restArgs) =>
+              dispatch(actionFunction({
+                ...idObj,
+                ...parentObj,
+                ...typeof obj === 'object' ? obj : { byKey: obj },
+              }, ...restArgs)
+            );
+          } else { 
+              dispatchableAction = (...args) => dispatch(actionFunction(...args));
           }
           
-          const name = stripped ? strippedName : fullName;
-          if (!strippedName.startsWith('clear')) {
-            dispatchableAction.isLoading = obj[`${name}IsLoading`]
-            dispatchableAction.error = obj[`${name}Error`]
-          }
-          return (
+          dispatchableAction.isLoading = obj[`${actionName}IsLoading`];
+          dispatchableAction.error = obj[`${actionName}Error`];
+          return { ...o, [actionName]: dispatchableAction };
+        }, {}),
+      ...Object.entries(asyncActionsIncluded).reduce(
+        (o, [actionName, actionFunction]) => {
+          const dispatchableAction = (...args) => dispatch(actionFunction(...args));
+          dispatchableAction.isLoading = obj[`${actionName}IsLoading`];
+          dispatchableAction.error = obj[`${actionName}Error`];
+          return { ...o, [actionName]: dispatchableAction };
+        }, {}),
+      ...Object.entries({ ...syncActionsIncluded, ...stripped ? syncActionsStripped : syncActions })
+          .reduce((o, [actionName, actionFunction]) => (
             {
               ...o,
-              [name]: dispatchableAction,
+              [actionName]: (...args) => dispatch(actionFunction(...args)),
             }
-          )
-          }, {}),
+          ), {}),
     }
   }