+++ src/utils.js
@@ -35,6 +35,24 @@ export const pluralToSingle = str => {
   } 
 }
 
+export const singleToPlural = str => {
+  if (str.slice(-1).toLowerCase() === 's') {
+    // This string is already plural: keep it unaltered
+    return str;
+  } else if (str.slice(-1) === 'y') {
+    // Handle special case of categories
+    return `${str.slice(0, -1)}ies`;
+  } else if (str.slice(-1) === 'Y') {
+    // Same but in upper case
+    return `${str.slice(0, -1)}IES`;
+  } else {
+    // Standard plural
+    return `${str}s`;
+  } 
+}
+
+export const titleCase = str => `${str.charAt(0).toUpperCase()}${str.slice(1)}`;
+
 // Simply call func with params if it is a function, if not ignore
 export const callIfFunc = (func, ...params) => {
   if (typeof func === 'function') {